chore(host): move random_get override
